app/pyproject.toml
[tool.poetry]
name = "take-home"
version = "0.1.0"
description = "Basic Django web application for take home assignment."
authors = ["Iván Abregú <ivan.abregu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.3"
psycopg2-binary = "^2.9.6"
requests = "^2.31.0"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
